+++ index.js
@@ -1,5 +1,5 @@
 var next = require('./lib/index.js');
 
-next.version = '1.22.19';
+next.version = '1.22.24';
 
 module.exports = next;